requirements.txt
appdirs==1.4.4
asgiref==3.3.4
attrs==21.2.0
autoflake==1.4
black==21.6b0
click==8.0.1
Django==3.2.4
djangorestframework==3.12.4
flake8==3.9.2
iniconfig==1.1.1
invoke==1.5.0
isort==5.9.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
regex==2021.4.4
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
